CM-107: removed LocalStorage and react-redux-router from @core and @login
Reverted back missing selectPoll action into redux/voting
Moved login handlers (now main method is onSubmitLoginForm)
Splitted navigation actions into navigation.js files (two of them)
Moved browser's history out of @core/@login
Incremented version of NPM packages: ethereumjs-wallet, web3 and web3-eth-accounts
Removed some file from core-dependencies
