--- packages/core-dependencies/router.js
@@ -1,8 +1,0 @@
-/**
- * Copyright 2017–2018, LaborX PTY
- * Licensed under the AGPL Version 3 license.
- *
- * @flow
- */
-
-export * from 'react-router-redux'
