+++ packages/login/package.json
@@ -11,7 +11,7 @@
     "bitcoinjs-lib": "github:chronobank/bitcoinjs-lib#opt-in-bitcoincash-sighash",
     "coinselect": "^3.1.11",
     "eth-crypto": "^1.2.0",
-    "ethereumjs-wallet": "0.6.0",
+    "ethereumjs-wallet": "0.6.2",
     "immutable": ">=3",
     "keythereum": "^1.0.4",
     "ledger-wallet-provider": "github:mikefluff/ledger-wallet-provider",
@@ -20,8 +20,8 @@
     "promisify-node-callback": "^1.0.2",
     "sockjs-client": "1.0.3",
     "uport-connect": "^0.7.3",
-    "web3": "1.0.0-beta.33",
+    "web3": "1.0.0-beta.35",
     "webstomp-client": "^1.2.0",
-    "web3-eth-accounts": "^1.0.0-beta.33"
+    "web3-eth-accounts": "^1.0.0-beta.35"
   }
 }
