+++ packages/login/redux/network/thunks.js
@@ -7,9 +7,6 @@
 
 import * as PersistAccountActions from '@chronobank/core/redux/persistAccount/actions'
 import {
-  createNetworkSession,
-  getProviderSettings,
-  login,
   selectProvider,
 } from '@chronobank/core/redux/session/thunks'
 import {
@@ -22,17 +19,12 @@ import {
 } from './constants'
 import * as NetworkActions from './actions'
 import uportProvider from '../../network/uportProvider'
-import privateKeyProvider from '../../network/privateKeyProvider'
-import walletProvider from '../../network/walletProvider'
-import setup from '../../network/EngineUtils'
 import web3Provider from '../../network/Web3Provider'
 import {
   getNetworkById,
   getNetworksByProvider,
-  LOCAL_PRIVATE_KEYS,
   NETWORK_MAIN_ID,
 } from '../../network/settings'
-import { DUCK_ETH_MULTISIG_WALLET } from '@chronobank/core/redux/multisigWallet/constants'
 
 /*
  * Thunk dispatched by "" screen.
@@ -97,37 +89,6 @@ export const initAccountsSignature = () =>
     dispatch(NetworkActions.resetLoadingAccountsSignatures())
   }
 
-export const handleWalletLogin = (wallet, password) => async (dispatch, getState) => {
-  dispatch(NetworkActions.loading())
-  dispatch(NetworkActions.clearErrors())
-
-  const providerSettings = dispatch(getProviderSettings())
-  const provider = walletProvider.getProvider(
-    wallet[0],
-    password,
-    providerSettings,
-  )
-
-  dispatch(NetworkActions.selectAccount(provider.ethereum.getAddress()))
-  await setup(provider)
-
-  const state = getState()
-  const {
-    selectedAccount,
-    selectedProviderId,
-    selectedNetworkId,
-  } = state.get(DUCK_NETWORK)
-
-  dispatch(NetworkActions.clearErrors())
-
-  dispatch(createNetworkSession(
-    selectedAccount,
-    selectedProviderId,
-    selectedNetworkId,
-  ))
-  await dispatch(login(selectedAccount))
-}
-
 /*
  * Thunk dispatched by "" screen.
  * TODO: to add description
@@ -136,43 +97,6 @@ export const initRecoverAccountPage = () => (dispatch) => {
   dispatch(NetworkActions.networkResetNewMnemonic())
   dispatch(NetworkActions.networkSetAccountRecoveryMode())
 }
-
-/*
- * Thunk dispatched by "" screen.
- * TODO: to add description
- * TODO: to rework it
- */
-export const handleLoginLocalAccountClick = (account = '') =>
-  async (dispatch, getState) => {
-    let state = getState()
-    const { accounts } = state.get(DUCK_NETWORK)
-    const wallets = state.get(DUCK_ETH_MULTISIG_WALLET)
-    const providerSetting = dispatch(getProviderSettings())
-    const index = Math.max(accounts.indexOf(account), 0)
-    const provider = privateKeyProvider.getPrivateKeyProvider(
-      LOCAL_PRIVATE_KEYS[index],
-      providerSetting,
-      wallets,
-    )
-    dispatch(NetworkActions.selectAccount(account))
-    await setup(provider)
-
-    state = getState()
-    const {
-      selectedAccount,
-      selectedProviderId,
-      selectedNetworkId,
-    } = state.get(DUCK_NETWORK)
-
-    dispatch(NetworkActions.clearErrors())
-
-    dispatch(createNetworkSession(
-      selectedAccount,
-      selectedProviderId,
-      selectedNetworkId,
-    ))
-    dispatch(login(selectedAccount))
-  }
 
 /*
  * Thunk dispatched by "" screen.
