+++ src/components/wallet/AddWalletWidget/AddWalletWidget.jsx
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { change, formValueSelector } from 'redux-form/immutable'
-import { resetWalletsForm } from '@chronobank/core/redux/mainWallet/actions'
+import { resetWalletsForm } from 'redux/ui/thunks'
 import { navigateToWallets } from 'redux/ui/navigation'
 import { FORM_ADD_NEW_WALLET } from '@chronobank/core/redux/mainWallet/constants'
 import WidgetContainer from 'components/WidgetContainer/WidgetContainer'
