+++ src/components/wallet/AddWalletWidget/CustomWalletForm/CustomWalletForm.jsx
@@ -11,7 +11,7 @@ import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
 import { TextField } from 'redux-form-material-ui'
 import { Field, formPropTypes, formValueSelector, reduxForm } from 'redux-form/immutable'
-import { resetWalletsForm } from '@chronobank/core/redux/mainWallet/actions'
+import { resetWalletsForm } from 'redux/ui/thunks'
 import { navigateToWallets } from 'redux/ui/navigation'
 import { createNewChildAddress } from '@chronobank/core/redux/wallets/actions'
 import { getChronobankTokens } from '@chronobank/core/redux/settings/erc20/tokens/selectors'
