+++ src/components/wallet/AddWalletWidget/SelectEthWallet/SelectEthWallet.jsx
@@ -8,7 +8,7 @@ import classnames from 'classnames'
 import { connect } from 'react-redux'
 import React, { PureComponent } from 'react'
 import { Translate } from 'react-redux-i18n'
-import { resetWalletsForm } from '@chronobank/core/redux/mainWallet/actions'
+import { resetWalletsForm } from 'redux/ui/thunks'
 import { navigateToWallets } from 'redux/ui/navigation'
 import { createNewChildAddress } from '@chronobank/core/redux/wallets/actions'
 import { BLOCKCHAIN_ETHEREUM } from '@chronobank/core/dao/constants'
