+++ src/components/wallet/AddWalletWidget/SelectWalletType/SelectWalletType.jsx
@@ -10,7 +10,7 @@ import classnames from 'classnames'
 import { Translate } from 'react-redux-i18n'
 import IPFSImage from 'components/common/IPFSImage/IPFSImage'
 import { TOKEN_ICONS } from 'assets'
-import { resetWalletsForm } from '@chronobank/core/redux/mainWallet/actions'
+import { resetWalletsForm } from 'redux/ui/thunks'
 import { navigateToWallets } from 'redux/ui/navigation'
 import { createNewChildAddress } from '@chronobank/core/redux/wallets/actions'
 import {
