+++ src/components/wallet/AddWalletWidget/StandardWalletForm/StandardWalletForm.jsx
@@ -10,7 +10,7 @@ import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
 import { TextField } from 'redux-form-material-ui'
 import { Field, formPropTypes, reduxForm } from 'redux-form/immutable'
-import { resetWalletsForm } from '@chronobank/core/redux/mainWallet/actions'
+import { resetWalletsForm } from 'redux/ui/thunks'
 import { navigateToWallets } from 'redux/ui/navigation'
 import { createNewChildAddress } from '@chronobank/core/redux/wallets/actions'
 import { FORM_CUSTOM_WALLET_ADD } from 'components/constants'
