+++ src/redux/configureStore.js
@@ -16,7 +16,7 @@ import { I18n, i18nReducer, loadTranslations, setLocale } from '@chronobank/core
 import moment from 'moment'
 import saveAccountMiddleWare from '@chronobank/core/redux/session/saveAccountMiddleWare'
 import thunk from 'redux-thunk'
-import LocalStorage from '@chronobank/core-dependencies/utils/LocalStorage'
+import LocalStorage from 'utils/LocalStorage'
 import coreReducers from '@chronobank/core/redux/ducks'
 import loginReducers from '@chronobank/login/redux/ducks'
 import { DUCK_I18N } from 'redux/i18n/constants'
